revert file rename
