+++ packages/client/hmi-client/src/workflow/ops/calibrate-ciemss/tera-calibrate-node-ciemss.vue
@@ -107,7 +107,7 @@
 
 <script setup lang="ts">
 import { computed, shallowRef, watch, ref, ComputedRef, onMounted, onUnmounted } from 'vue';
-import { ProgressState, WorkflowNode } from '@/types/operator';
+import { ProgressState, WorkflowNode } from '@/types/workflow';
 import DataTable from 'primevue/datatable';
 import Button from 'primevue/button';
 import Dropdown from 'primevue/dropdown';
