+++ packages/client/hmi-client/src/workflow/ops/calibrate-julia/tera-calibrate-node-julia.vue
@@ -104,7 +104,7 @@
 
 <script setup lang="ts">
 import { computed, shallowRef, watch, ref, ComputedRef, onMounted, onUnmounted } from 'vue';
-import { ProgressState, WorkflowNode } from '@/types/operator';
+import { ProgressState, WorkflowNode } from '@/types/workflow';
 import DataTable from 'primevue/datatable';
 import Button from 'primevue/button';
 import Column from 'primevue/column';
