+++ packages/client/hmi-client/src/workflow/ops/dataset-transformer/tera-dataset-transformer-node.vue
@@ -6,7 +6,7 @@
 
 <script setup lang="ts">
 import { watch } from 'vue';
-import { WorkflowNode, WorkflowPortStatus } from '@/types/operator';
+import { WorkflowNode, WorkflowPortStatus } from '@/types/workflow';
 import { DatasetTransformerState } from './dataset-transformer-operation';
 
 const props = defineProps<{
