+++ packages/client/hmi-client/src/workflow/ops/dataset-transformer/tera-dataset-transformer.vue
@@ -15,7 +15,7 @@
 <script setup lang="ts">
 // Proxy to use tera-dataset via a workflow context
 
-import { WorkflowNode, WorkflowPortStatus } from '@/types/operator';
+import { WorkflowNode, WorkflowPortStatus } from '@/types/workflow';
 import TeraDatasetJupyterPanel from '@/components/dataset/tera-dataset-jupyter-panel.vue';
 import { computed, onMounted, ref } from 'vue';
 import { workflowEventBus } from '@/services/workflow';
