+++ packages/client/hmi-client/src/workflow/ops/dataset/tera-dataset-workflow-wrapper.vue
@@ -5,7 +5,7 @@
 <script setup lang="ts">
 // Proxy to use tera-dataset via a workflow context
 
-import { WorkflowNode } from '@/types/operator';
+import { WorkflowNode } from '@/types/workflow';
 import TeraDataset from '@/components/dataset/tera-dataset.vue';
 import { DatasetOperationState } from './dataset-operation';
