+++ packages/client/hmi-client/src/workflow/ops/funman/tera-funman.vue
@@ -131,7 +131,7 @@ import Dropdown from 'primevue/dropdown';
 import Slider from 'primevue/slider';
 import { FunmanPostQueriesRequest, Model, ModelConfiguration } from '@/types/Types';
 import { getQueries, makeQueries } from '@/services/models/funman-service';
-import { WorkflowNode } from '@/types/operator';
+import { WorkflowNode } from '@/types/workflow';
 import { workflowEventBus } from '@/services/workflow';
 import teraConstraintGroupForm from '@/components/funman/tera-constraint-group-form.vue';
 import teraFunmanOutput from '@/components/funman/tera-funman-output.vue';
