+++ packages/client/hmi-client/src/workflow/ops/model-from-code/model-from-code-operation.ts
@@ -1,4 +1,4 @@
-import { Operation, WorkflowOperationTypes } from '@/types/operator';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 
 export interface ModelFromCodeState {
 	codeLanguage: string;
