+++ packages/client/hmi-client/src/workflow/ops/model-transformer/model-transformer-operation.ts
@@ -1,4 +1,4 @@
-import { Operation, WorkflowOperationTypes } from '@/types/operator';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 
 export interface ModelTransformerState {
 	modelId: string | null;
