+++ packages/client/hmi-client/src/workflow/ops/model-transformer/tera-model-transformer-node.vue
@@ -9,7 +9,7 @@
 <script setup lang="ts">
 import { getModel } from '@/services/model';
 import { Model } from '@/types/Types';
-import { WorkflowNode } from '@/types/operator';
+import { WorkflowNode } from '@/types/workflow';
 import { onMounted, ref, watch } from 'vue';
 import TeraModelDiagram from '@/components/model/petrinet/model-diagrams/tera-model-diagram.vue';
 import { ModelTransformerState } from './model-transformer-operation';
