+++ packages/client/hmi-client/src/workflow/ops/simulate-ciemss/simulate-ciemss-operation.ts
@@ -1,5 +1,5 @@
 import { TimeSpan } from '@/types/Types';
-import { Operation, WorkflowOperationTypes } from '@/types/operator';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 import { SimulationConfig } from '@/types/SimulateConfig';
 
 export interface SimulateCiemssOperationState {
