+++ packages/client/hmi-client/src/workflow/ops/simulate-ciemss/tera-simulate-node-ciemss.vue
@@ -63,7 +63,7 @@ import {
 	getSimulation
 } from '@/services/models/simulation-service';
 import InputNumber from 'primevue/inputnumber';
-import { ProgressState, WorkflowNode } from '@/types/operator';
+import { ProgressState, WorkflowNode } from '@/types/workflow';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
 import { workflowEventBus } from '@/services/workflow';
 import { Poller, PollerState } from '@/api/api';
