+++ packages/client/hmi-client/src/workflow/ops/simulate-ensemble-ciemss/simulate-ensemble-ciemss-operation.ts
@@ -1,4 +1,4 @@
-import { Operation, WorkflowOperationTypes } from '@/types/operator';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 // import { EnsembleRequest } from '@/types/Types';
 // import { makeEnsembleJob } from '@/services/models/simulation-service';
 import { ChartConfig } from '@/types/SimulateConfig';
