+++ packages/client/hmi-client/src/workflow/ops/stratify-mira/stratify-mira-operation.ts
@@ -1,4 +1,4 @@
-import { Operation, WorkflowOperationTypes } from '@/types/operator';
+import { Operation, WorkflowOperationTypes } from '@/types/workflow';
 
 export interface StratifyGroup {
 	borderColour: string;
